# Update OptimizedSolver