--- a/pentomino-solver-cli/src/bin/main.rs
+++ b/pentomino-solver-cli/src/bin/main.rs
@@ -87,13 +87,13 @@ fn main() {
     let args = Args::parse();
 
     let ((rows, cols), initial) = match args.board {
-        Board::Rect3x20 => ((3, 20), 0.into()),
-        Board::Rect4x15 => ((4, 15), 0.into()),
-        Board::Rect5x12 => ((5, 12), 0.into()),
-        Board::Rect6x10 => ((6, 10), 0.into()),
+        Board::Rect3x20 => ((3, 20), 0),
+        Board::Rect4x15 => ((4, 15), 0),
+        Board::Rect5x12 => ((5, 12), 0),
+        Board::Rect6x10 => ((6, 10), 0),
         Board::Rect8x8_2x2 => (
             (8, 8),
-            [27, 28, 35, 36].iter().map(|&p| 1 << p).sum::<u64>().into(),
+            [27, 28, 35, 36].iter().map(|&p| 1 << p).sum::<u64>(),
         ),
     };
     let solver = args.solver.create_solver(rows, cols);