--- a/pentomino-solver/src/lib.rs
+++ b/pentomino-solver/src/lib.rs
@@ -1,8 +1,8 @@
-mod bitboard;
 mod pieces;
 mod shapes;
 pub mod solvers;
 
-pub use bitboard::Bitboard;
 pub use pieces::{Piece, NUM_PIECES};
 pub use solvers::Solver;
+
+type Bitboard = u64;