--- a/pentomino-solver/src/lib.rs
+++ b/pentomino-solver/src/lib.rs
@@ -5,4 +5,4 @@ pub mod solvers;
 pub use pieces::{Piece, NUM_PIECES};
 pub use solvers::Solver;
 
-type Bitboard = u64;
+pub type Bitboard = u64;