--- a/pentomino-solver/src/solvers.rs
+++ b/pentomino-solver/src/solvers.rs
@@ -4,7 +4,7 @@ mod simple;
 
 use crate::{Bitboard, Piece, NUM_PIECES};
 pub use default::DefaultSolver;
-pub use optimized::OptimizedSolver;
+pub use optimized::{OptimizedSolver, OptimizedSolverType};
 pub use simple::SimpleSolver;
 
 pub trait Solver {
@@ -14,5 +14,5 @@ pub trait Solver {
 
 pub(crate) trait SolutionStore {
     fn add_solution(&mut self, pieces: &[Bitboard; NUM_PIECES]);
-    fn get_solutions(self) -> Vec<[Bitboard; NUM_PIECES]>;
+    fn get_solutions(&self) -> Vec<[Bitboard; NUM_PIECES]>;
 }