… pentomino-solver/src/solvers/simple.rs