before: #4 

after:
```
test bench_3x20_optimized           ... bench:     194,486 ns/iter (+/- 21,178)
test bench_3x20_optimized_unique    ... bench:     189,127 ns/iter (+/- 3,525)
test bench_4x15_optimized           ... bench:   4,161,670 ns/iter (+/- 822,803)
test bench_4x15_optimized_unique    ... bench:   4,224,275 ns/iter (+/- 315,255)
test bench_5x12_optimized           ... bench:  23,031,266 ns/iter (+/- 990,321)
test bench_5x12_optimized_unique    ... bench:  23,568,891 ns/iter (+/- 2,681,557)
test bench_6x10_optimized           ... bench:  57,709,916 ns/iter (+/- 3,398,636)
test bench_6x10_optimized_unique    ... bench:  58,370,725 ns/iter (+/- 2,891,794)
test bench_8x8_2x2_optimized        ... bench:  13,767,500 ns/iter (+/- 1,390,155)
test bench_8x8_2x2_optimized_unique ... bench:  14,012,054 ns/iter (+/- 1,266,752)
```