@sugyan 2023-09-28
```
test bench_3x20_optimized           ... bench:     195,157 ns/iter (+/- 1,905)
test bench_3x20_optimized_unique    ... bench:     156,276 ns/iter (+/- 15,663)
test bench_4x15_optimized           ... bench:   4,081,533 ns/iter (+/- 447,083)
test bench_4x15_optimized_unique    ... bench:   4,026,929 ns/iter (+/- 183,525)
test bench_5x12_optimized           ... bench:  21,460,554 ns/iter (+/- 738,862)
test bench_5x12_optimized_unique    ... bench:  21,514,058 ns/iter (+/- 3,374,037)
test bench_6x10_optimized           ... bench:  53,655,425 ns/iter (+/- 629,025)
test bench_6x10_optimized_unique    ... bench:  53,449,541 ns/iter (+/- 3,309,326)
test bench_8x8_2x2_optimized        ... bench:  11,335,779 ns/iter (+/- 692,667)
test bench_8x8_2x2_optimized_unique ... bench:  11,470,558 ns/iter (+/- 1,311,264)
```