@sugyan 2023-09-29
```
test bench_3x20_optimized           ... bench:     106,988 ns/iter (+/- 2,569)
test bench_3x20_optimized_unique    ... bench:     106,892 ns/iter (+/- 3,436)
test bench_4x15_optimized           ... bench:   3,790,470 ns/iter (+/- 17,543)
test bench_4x15_optimized_unique    ... bench:   3,682,766 ns/iter (+/- 446,656)
test bench_5x12_optimized           ... bench:  20,925,983 ns/iter (+/- 2,938,451)
test bench_5x12_optimized_unique    ... bench:  20,900,437 ns/iter (+/- 2,845,677)
test bench_6x10_optimized           ... bench:  51,175,475 ns/iter (+/- 2,701,502)
test bench_6x10_optimized_unique    ... bench:  50,125,108 ns/iter (+/- 2,312,633)
test bench_8x8_2x2_optimized        ... bench:  10,483,337 ns/iter (+/- 349,387)
test bench_8x8_2x2_optimized_unique ... bench:  10,338,504 ns/iter (+/- 1,035,515)
```