@sugyan 2023-10-03
```
test bench_3x20_optimized_large           ... bench:     186,460 ns/iter (+/- 13,175)
test bench_3x20_optimized_large_unique    ... bench:     189,662 ns/iter (+/- 14,860)
test bench_3x20_optimized_small           ... bench:     291,683 ns/iter (+/- 12,010)
test bench_3x20_optimized_small_unique    ... bench:     291,343 ns/iter (+/- 7,627)
test bench_4x15_optimized_large           ... bench:   3,838,033 ns/iter (+/- 1,037,680)
test bench_4x15_optimized_large_unique    ... bench:   3,823,687 ns/iter (+/- 867,478)
test bench_4x15_optimized_small           ... bench:   4,792,504 ns/iter (+/- 125,497)
test bench_4x15_optimized_small_unique    ... bench:   4,762,575 ns/iter (+/- 183,179)
test bench_5x12_optimized_large           ... bench:  19,996,983 ns/iter (+/- 3,243,306)
test bench_5x12_optimized_large_unique    ... bench:  19,773,566 ns/iter (+/- 1,744,545)
test bench_5x12_optimized_small           ... bench:  24,129,850 ns/iter (+/- 388,582)
test bench_5x12_optimized_small_unique    ... bench:  24,206,195 ns/iter (+/- 1,760,916)
test bench_6x10_optimized_large           ... bench:  50,567,541 ns/iter (+/- 3,245,711)
test bench_6x10_optimized_large_unique    ... bench:  49,866,116 ns/iter (+/- 1,745,987)
test bench_6x10_optimized_small           ... bench:  60,824,408 ns/iter (+/- 586,787)
test bench_6x10_optimized_small_unique    ... bench:  60,761,958 ns/iter (+/- 2,162,156)
test bench_8x8_2x2_optimized_large        ... bench:  10,392,233 ns/iter (+/- 1,068,640)
test bench_8x8_2x2_optimized_large_unique ... bench:  10,552,325 ns/iter (+/- 1,148,723)
test bench_8x8_2x2_optimized_small        ... bench:  12,694,300 ns/iter (+/- 427,628)
test bench_8x8_2x2_optimized_small_unique ... bench:  12,678,208 ns/iter (+/- 404,111)
```