@sugyan 2023-10-04
```
test bench_3x20_optimized_large           ... bench:     146,794 ns/iter (+/- 3,932)
test bench_3x20_optimized_large_unique    ... bench:     148,133 ns/iter (+/- 11,285)
test bench_3x20_optimized_small           ... bench:     222,727 ns/iter (+/- 4,791)
test bench_3x20_optimized_small_unique    ... bench:     222,259 ns/iter (+/- 7,741)
test bench_4x15_optimized_large           ... bench:   3,423,866 ns/iter (+/- 163,258)
test bench_4x15_optimized_large_unique    ... bench:   3,364,483 ns/iter (+/- 228,070)
test bench_4x15_optimized_small           ... bench:   4,224,250 ns/iter (+/- 104,524)
test bench_4x15_optimized_small_unique    ... bench:   4,176,050 ns/iter (+/- 173,276)
test bench_5x12_optimized_large           ... bench:  17,776,679 ns/iter (+/- 1,058,340)
test bench_5x12_optimized_large_unique    ... bench:  17,583,633 ns/iter (+/- 921,629)
test bench_5x12_optimized_small           ... bench:  21,223,533 ns/iter (+/- 374,548)
test bench_5x12_optimized_small_unique    ... bench:  21,077,758 ns/iter (+/- 440,215)
test bench_6x10_optimized_large           ... bench:  45,576,979 ns/iter (+/- 9,106,196)
test bench_6x10_optimized_large_unique    ... bench:  43,768,287 ns/iter (+/- 3,552,598)
test bench_6x10_optimized_small           ... bench:  53,389,466 ns/iter (+/- 1,215,779)
test bench_6x10_optimized_small_unique    ... bench:  53,038,533 ns/iter (+/- 643,387)
test bench_8x8_2x2_optimized_large        ... bench:   9,218,137 ns/iter (+/- 1,760,003)
test bench_8x8_2x2_optimized_large_unique ... bench:   9,098,916 ns/iter (+/- 1,236,166)
test bench_8x8_2x2_optimized_small        ... bench:  11,231,212 ns/iter (+/- 743,695)
test bench_8x8_2x2_optimized_small_unique ... bench:  11,177,375 ns/iter (+/- 266,481)
```